commit 03/12

+++ gr02/Workspace/audio_filter/fi_custom.c
@@ -1,13 +1,13 @@
 #include "fi_custom.h"
 
-int mpy_fi(int A, int B, int y) {
+long int mpy_fi(int A, int B, int y) {
 
 	long int C = (long)B * A;
 	long int add = 0x1 << y - 1;
 	C = C + add;
 	C = C >> y;
 
-	return C;
+	return (long)C;
 }
 
 //ADD
